DB연결 및 코드 수정

+++ src/main/java/com/cloud/cloudproject/RequestItem/RequestItemDTO.java
@@ -1,21 +1,18 @@
 package com.cloud.cloudproject.RequestItem;
 
 import lombok.*;
-
 import java.util.Date;
-
 
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class RequestItemDTO {
-    private int postId;
+    private Long postId;  // int → Long 변경
     private String title;
     private String content;
     private Date date;
     private String categoryCode;
-    private int userCode;
-    private int startCash;
-
+    private Integer userCode;  // int → Integer 변경 (null 허용)
+    private Integer startCash;  // int → Integer 변경 (null 허용)
 }
